Update quiz generation
