+++ README.md
@@ -4,4 +4,4 @@ Generate quizes in a particular field.
 Intall required dependencies
 Add your env variables
 
-Run app with `streamlit run quiz-app.py`
+Run app with `streamlit run quiz.py`
